SCRUM-3121 created TaxonFormEditor
