+++ src/main/cliapp/src/containers/allelesPage/AlleleDetailPage.js
@@ -2,11 +2,20 @@ import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useQuery } from 'react-query';
 import { AlleleService } from '../../service/AlleleService';
+import ErrorBoundary from '../../components/Error/ErrorBoundary';
+import { TaxonAdditionalFieldData } from '../../components/FieldData/TaxonAdditionalFieldData';
+import { FormErrorMessageComponent } from '../../components/Error/FormErrorMessageComponent';
+import { TaxonFormEditor } from '../../components/Editors/taxon/TaxonFormEditor';
 
 export default function AlleleDetailPage(){
 	const [allele, setAllele] = useState();
 	const { curie } = useParams();
 	const alleleService = new AlleleService();
+	const errorMessages = useState([]);
+
+	const labelColumnSize = "col-3";
+	const widgetColumnSize = "col-4";
+	const fieldDetailsColumnSize = "col-5";
 
 	useQuery([curie],
 		() => alleleService.getAllele(curie), 
@@ -22,25 +31,29 @@ export default function AlleleDetailPage(){
 		}
 	);
 
+	const onTaxonValueChange = () => {
+		console.log("onTaxonValueChange");
+	}
+
 	return(
 		<>
 			<h1>Allele Detail Page</h1>
-				<ErrorBoundary>
+			<ErrorBoundary>
 				<form>
 					<div className="grid">
 						<div className={labelColumnSize}>
-					 		<label htmlFor="subject"><font color={'red'}>*</font>Subject</label>
+					 		<label htmlFor="taxon"><font color={'red'}>*</font>Taxon</label>
 						</div>
 						<div className={widgetColumnSize}>
+							<TaxonFormEditor curie={allele?.taxon?.curie} onTaxonValueChange={onTaxonValueChange} />
 						</div>
 						<div className={fieldDetailsColumnSize}>
-							<FormErrorMessageComponent errorMessages={errorMessages} errorField={"subject"}/>
-							<FormErrorMessageComponent errorMessages={uiErrorMessages} errorField={"subject"}/>
-							<SubjectAdditionalFieldData fieldData={newAnnotation.subject}/>
+							<FormErrorMessageComponent errorMessages={errorMessages} errorField={"taxon"}/>
+							<TaxonAdditionalFieldData curie={allele?.taxon?.curie}/>
 						</div>
 					</div>
 
-					<div className="grid">
+					{/* <div className="grid">
 						<div className={labelColumnSize}>
 							<label htmlFor="assertedGenes">Asserted Genes</label>
 						</div>
@@ -97,10 +110,10 @@ export default function AlleleDetailPage(){
 							<FormErrorMessageComponent errorMessages={errorMessages} errorField={"object"}/>
 							<DiseaseAdditionalFieldData fieldData={newAnnotation.object}/>
 						</div>
-					</div>
+					</div> */}
 
 			</form>
-				</ErrorBoundary>
+		</ErrorBoundary>
 		</>
 	)
